+++ rprpp/Error.cpp
@@ -1,5 +1,5 @@
 #include "Error.h"
-#include "vkdx11_interop/rprpp.h"
+#include "vkdx11_interop/vkdx11_interop.h"
 
 namespace rprpp {
