+++ vkdx11_interop/vkdx11_interop.cpp
@@ -1,4 +1,4 @@
-#include "rprpp.h"
+#include "vkdx11_interop.h"
 
 #include "rprpp/Error.h"
 #include "rprpp/PostProcessing.h"
